+++ src/main/java/com/api/ttoklip/domain/profile/domain/ProfileRepository.java
@@ -1,0 +1,5 @@
+package com.api.ttoklip.domain.profile.domain;
+
+public interface ProfileRepository {
+    void save(Profile profile);
+}
