+++ src/main/java/com/api/ttoklip/global/security/auth/service/AuthService.java
@@ -7,9 +7,9 @@ import com.api.ttoklip.domain.member.domain.vo.Role;
 import com.api.ttoklip.domain.member.repository.MemberRepository;
 import com.api.ttoklip.domain.member.service.MemberService;
 import com.api.ttoklip.domain.privacy.domain.Interest;
-import com.api.ttoklip.domain.privacy.domain.Profile;
+import com.api.ttoklip.domain.profile.domain.Profile;
 import com.api.ttoklip.domain.privacy.repository.InterestRepository;
-import com.api.ttoklip.domain.privacy.service.ProfileService;
+import com.api.ttoklip.domain.profile.service.ProfileService;
 import com.api.ttoklip.domain.term.domain.Term;
 import com.api.ttoklip.domain.term.domain.TermAgreement;
 import com.api.ttoklip.domain.term.service.TermService;
