+++ src/main/java/com/api/ttoklip/global/security/oauth2/service/OAuthRegisterService.java
@@ -4,8 +4,8 @@ import com.api.ttoklip.domain.member.domain.Member;
 import com.api.ttoklip.domain.member.domain.vo.Provider;
 import com.api.ttoklip.domain.member.domain.vo.Role;
 import com.api.ttoklip.domain.member.service.MemberService;
-import com.api.ttoklip.domain.privacy.domain.Profile;
-import com.api.ttoklip.domain.privacy.service.ProfileService;
+import com.api.ttoklip.domain.profile.domain.Profile;
+import com.api.ttoklip.domain.profile.service.ProfileService;
 import com.api.ttoklip.global.security.oauth2.userInfo.OAuth2UserInfo;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
