+++ src/testFixtures/java/member/fixture/MemberFixture.java
@@ -2,7 +2,7 @@ package member.fixture;
 
 import com.api.ttoklip.domain.member.domain.Member;
 import com.api.ttoklip.domain.member.domain.vo.Role;
-import com.api.ttoklip.domain.privacy.domain.Profile;
+import com.api.ttoklip.domain.profile.domain.Profile;
 import java.util.List;
 
 @SuppressWarnings("NonAsciiCharacters")
